Add gRPC greeter sample application
This sample application consists of a gRPC server and client.

It is a adapted from the gRPC-Go helloworld example:
https://github.com/grpc/grpc-go/tree/master/examples/helloworld

The main changes from the original version are:

- the server returns the hostname where it runs in a gRPC header; and

- the client prints the hostname header from the gRPC response.

This is useful for testing request routing and load balancing
functionality.

When running in a Kubernetes cluster, the hostname is the name of the
Kubernetes Pod that served the request.